program that finds and prints the first 98 Fibonacci numbers, starting with 1 and 2

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,7 +9,7 @@
 int main(void)
 {
 	int i;
-	unsigned int fib, prv_fib1, prv_fib2;
+	unsigned long int fib, prv_fib1, prv_fib2;
 
 	prv_fib1 = 0;
 	prv_fib2 = 1;
